backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.11.5
pydantic[email]
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.12
llama-index==0.10.0 
neo4j==5.15.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
httpx==0.27.0
python-dotenv==1.0.0
openai>=1.0.0